make clear that sleep represents integer (ms) not float (s)

+++ lib/conway_deathmatch/cli.ex
@@ -26,7 +26,7 @@ defmodule ConwayDeathmatch.CLI do
                                           width:  :integer,
                                           height: :integer,
                                           ticks:  :integer,
-                                          sleep:  :float,
+                                          sleep:  :integer,
                                           shapes: :string,
                                           step:   :boolean,
                                           one:    :string,
